requirements.txt
flask==1.1.2
mysql-connector-python==8.0.22